Hardcode UTF8 parsed string to avoid Mojibake
The conditional logic is to avoid needing to 'use utf8' which might barf
in the wrong place on earlier perls without configure_requires support.

This could also be done by simply placing the whole text inline on the
'dist_abstract' line and then adding 'use utf8' at the top of the file,
but this approach is slightly more portable.

And I'm not sure what versions of Module::Build support this trick, but
the most recent one on CPAN does.

Hopefully, this commit should resolve #1